Basic Inventory Structure
Basic Invenetory sctructure build

to Create a Armour Item, in the project, in the correct folder right-click, create>Inventory System>Items>"ItemType;
Name the new Item, assign values to the item.
go to the ItemDataBase, and Add the newly created item(A scriptableObject) Item to that Database.
From there create a new gameObject, attach the sprite we want, add a Box Collider component to it, set it to Trigger, and attach the item script to it, and drag the ScriptedObject we created earlier into the item scripts box. We need to created a Panel, call it inventory, then resize it to the size we want for the Inventory, Add the InventoryDisplay to that inventory, we then need to set the X start and Y Start relative to the size of the inventory panel, 0 is dead center, usualt -170 through -190 works, then the size of the spaces between slots, Set the number of Colums to 10 (For this game, increase for larger inventories, and set the y Spacing for a second row (My code has us locekd to 10 slots, but I can change it)
Take the BASE_ITEM_SLOT and put it in the editor, rename it to "NewItemName"_Slot, and change the images around, the "Top" layer is set to Emptyslot sprite we have, and the second image is set to the item we created, create a prefab variant of that itemslot, go back to the ScriptableObject item we created, and drag that new Prefab_slot into the prefab parameters in the inspector.

Makea prefab of the gameObject we attached the Item script to, thats the item in the game world. add as many as we need per scene.

We havea new object that can be placed in the inventory and stacked (Currently an unknown amount per stack)