Trying to get the Dias Working
Dias not working, should be as follows,
if the Dias is colliding with the Gem tagged object, the gem should dissapear(Have not implemented that part yet) set hasGem to true, and change the sprite to the active button sprte
but its not. hoping someone else can take a look and see what I missed.